backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
cryptography==41.0.7
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
alembic==1.12.1
fhirclient==4.1.0
httpx==0.25.1
pandas==2.1.3
numpy==1.26.2
openpyxl==3.1.2
pyarrow==14.0.1
redis==5.0.1
celery==5.3.4

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.1
faker==20.1.0

# Code quality tools
black==23.12.0
flake8==6.1.0
pylint==3.0.3
mypy==1.7.1
isort==5.13.2

